requirements.txt
aiohttp==3.12.14
chromadb==1.0.15
dataclasses-json==0.6.7
hdbscan==0.8.40
huggingface-hub==0.33.2
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-huggingface==0.3.0
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.5
numpy==2.3.1
openai==1.95.0
pandas==2.3.1
python-dotenv==1.1.1
scikit-learn==1.7.0
sentence-transformers==5.0.0
tqdm==4.67.1
uvicorn==0.35.0